pyproject.toml
[project]
name = "ucar-ingestion"
version = "0.1.0"
description = "RAG Ingestion Pipeline for UCAR ERP - Document parsing, chunking, vectorization"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "supabase>=2.7.0",
    "unstructured-client>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.7.0",
    "numpy>=1.26.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",
    "nats-py>=2.8.0",
    "structlog>=24.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]